# api/openapi.yaml
openapi: 3.1.0
info:
  title: BugBarn API
  description: |
    Self-hosted error tracking, structured logging, and web analytics.

    ## Authentication

    Most endpoints require authentication via one of two methods:

    - **Session cookie** — Log in via `POST /api/v1/login` to receive a `bugbarn_session` cookie.
      State-changing requests (POST, PUT, PATCH, DELETE) must include the `X-BugBarn-CSRF` header
      with the value from the `bugbarn_csrf` cookie.
    - **API key** — Pass a key in the `X-BugBarn-API-Key` header. Keys have scopes:
      `full` (all access), `read` (GET only, no settings/apikeys), or `ingest` (POST to ingest endpoints only).

    Ingest endpoints (`/api/v1/events`, `/api/v1/logs`, `/api/v1/analytics/collect`) use
    wildcard CORS and accept ingest-scoped API keys so browser SDKs can POST from any origin.
  version: 1.0.0
  contact:
    name: BugBarn
    url: https://bugbarn.dev
servers:
  - url: https://bugbarn.wiebe.xyz
    description: Production
  - url: http://localhost:8080
    description: Local development
...
security:
  - sessionCookie: []
  - apiKey: []
paths:
...
  /api/v1/oidc/backchannel-logout:
    post:
      operationId: oidcBackchannelLogout
      tags: [Auth]
      summary: OIDC Back-Channel Logout 1.0 endpoint
      description: |
        Called server-to-server by the identity provider when a user's IdP
        session ends (logout, admin revocation, suspension). The signed
        `logout_token` is verified against the issuer's JWKS; matching local
        sessions (by `sid`, else by `sub`) are deleted immediately.
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [logout_token]
              properties:
                logout_token:
                  type: string
                  description: Signed JWT logout token minted by the IdP
      responses:
        "200":
          description: Matching sessions terminated (also returned when none matched)
        "400":
          description: Invalid logout token
        "429":
          description: Rate limited
...
  /api/v1/releases:
    get:
      operationId: listReleases
      tags: [Releases]
      summary: List releases
      responses:
        "200":
          description: List of releases
          content:
            application/json:
              schema:
                type: object
                properties:
                  releases:
                    type: array
                    items:
                      $ref: "#/components/schemas/Release"
    post:
      operationId: createRelease
      tags: [Releases]
      summary: Create a release marker
      description: >
        Accepts a release marker and enqueues it for asynchronous creation by the
        background worker, returning 202 immediately. This keeps the request off
        the single SQLite writer connection. The marker is typically persisted
        within ~1s and becomes visible via GET /api/v1/releases shortly after.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReleaseInput"
      responses:
        "202":
          description: Release marker accepted for asynchronous creation
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted:
                    type: boolean
                  ingestId:
                    type: string
...
  /api/v1/settings:
    get:
      operationId: getSettings
      tags: [Settings]
      summary: Get project settings
      responses:
        "200":
          description: Settings map
          content:
            application/json:
              schema:
                type: object
                properties:
                  settings:
                    type: object
                    additionalProperties:
                      type: string
    put:
      operationId: updateSettings
      tags: [Settings]
      summary: Update project settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  settings:
                    type: object
                    additionalProperties:
                      type: string

  # ──────────────────────────────────────────────
  # Source Maps
  # ──────────────────────────────────────────────
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: bugbarn_session
      description: |
        Session cookie set by POST /api/v1/login. State-changing requests
        must also include the X-BugBarn-CSRF header with the value from
        the bugbarn_csrf cookie.
    apiKey:
      type: apiKey
      in: header
      name: X-BugBarn-API-Key
      description: |
        API key header. Scopes: `full` (all access), `read` (GET only, excludes
        settings and apikeys), `ingest` (POST to ingest endpoints only).
...
  schemas:
...
    Release:
      type: object
      properties:
        ID:
          type: string
        Name:
          type: string
        Environment:
          type: string
        ObservedAt:
          type: string
          format: date-time
        Version:
          type: string
        CommitSHA:
          type: string
        URL:
          type: string
        Notes:
          type: string
        CreatedBy:
          type: string
        CreatedAt:
          type: string
          format: date-time
    ReleaseInput:
      type: object
      properties:
        name:
          type: string
        environment:
          type: string
        observedAt:
          type: string
          format: date-time
        version:
          type: string
        commitSha:
          type: string
        url:
          type: string
        notes:
          type: string
        createdBy:
          type: string
...